refactor: separate FieldComponents
